abandoning state features

+++ src/Entity.h
@@ -26,7 +26,7 @@ public:
     bigint frame_progress;
 
 protected:
-    Entity(bool is_moving_entity, const Vector2 & center, const Vector2 & size, const sf::Color & color, float elasticity, float friction) :
+    Entity(bool is_moving_entity, const Vector2 & center, const Vector2 & size, const sf::Color & color, bigint elasticity, bigint friction) :
         is_moving_entity(is_moving_entity), center(center), size(size), color(color), elasticity(elasticity), friction(friction) {}
 
     static void makeRectangle(std::vector<Vector2>* polygon, const Vector2 &center, const Vector2 &size);
